Model rendered

+++ VulkanTutorial.LoadingModels/VulkanIndexBuffer.cs
@@ -2,12 +2,12 @@
 
 namespace VulkanTutorial.DepthBuffering;
 
-public class VulkanIndexBuffer : VulkanDeviceBuffer<ushort>
+public class VulkanIndexBuffer : VulkanDeviceBuffer<uint>
 {
     public VulkanIndexBuffer(Vk vk,
         VulkanVirtualDevice device, VulkanCommandPool commandPool,
-        ushort[] indices) : base(vk, device, commandPool, indices,
+        uint[] indices) : base(vk, device, commandPool, indices,
             BufferUsageFlags.BufferUsageIndexBufferBit) { }
 
-    public override void Bind(in CommandBuffer commandBuffer) => this.Vk.CmdBindIndexBuffer(commandBuffer, this.Buffer, 0, IndexType.Uint16);
+    public override void Bind(in CommandBuffer commandBuffer) => this.Vk.CmdBindIndexBuffer(commandBuffer, this.Buffer, 0, IndexType.Uint32);
 }
